Migración a Arquitectura Basada en API RESTful/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.3
python-dotenv==1.0.1
gunicorn==22.0.0
